+++ test/transform/resource/before/BuilderCustomName.java
@@ -1,0 +1,7 @@
+//CONF: lombok.builder.className = SimpleTestBuilder
+import java.util.List;
+
+@lombok.experimental.SuperBuilder
+class BuilderCustomName<T> {
+	private final int field;
+}